modules/osint.py
import requests
from bs4 import BeautifulSoup

def perform_google_dorks(target_url, api_key=None):
    """
    Performs basic Google Dork searches.
    Note: For a robust AI-driven approach, this would use a custom model to generate dorks.
    """
    dorks = [
        f"site:{target_url} filetype:pdf",
        f"site:{target_url} intitle:index.of",
        f"site:{target_url} inurl:admin",
        f"site:{target_url} 'password'",
        f"site:{target_url} 'API_KEY'"
    ]
    
    results = {}
    for dork in dorks:
        try:
            # Simple web scraping for a proof of concept.
            # A real-world tool would use Google's Custom Search API
            # to avoid being blocked.
            response = requests.get(f"https://www.google.com/search?q={dork}", headers={'User-Agent': 'Mozilla/5.0'})
            soup = BeautifulSoup(response.text, 'html.parser')
            # Look for search result links
            links = [a['href'] for a in soup.find_all('a', href=True) if a['href'].startswith('http') and target_url in a['href']]
            results[dork] = links
        except Exception as e:
            results[dork] = f"Error: {e}"
    
    return results