requirements.txt
requests
beautifulsoup4
python-nmap
pandas
reportlab # or another PDF library like fpdf